Fixed Title Screen Bug
I fixed the bug where since I added the title screen I couldn't just reset the scene, so I added a check to see if the game was just loading up or if it was after a death. Now it works as intended, where the title screen anim only runs when you first open the game, and not after every death.