Started Work on the other Two Bosses
Started to add the dragon and the airship. Decided I might as well get that part over with so that I can make a good trailer with all of them in. Might not add all the actual abilities and stuff, but at least need to have it all animated so I can film them.