Finished Aviary Anims and Demo tag
The Aviary open and close feature works now and the carousel works as designed. Still no functionality as you cannot buy anything, but thats the same with the upgrade store atm.

Also added a feature where I can change a variable named 'demo' true or false and it will automatically have my load my demo build or load the full game depending. This will make it easier to export for the demo and/or the full game simultaneously.