Added Tips
Added a customizable tip feature so that I can help out the player in certain spots where they might not understand. It now explains how to beat the boss, and when you first get hit, it will tell you that mosquitos replenish a little bit of HP.

Also cleaned up the SaverLoader a bit. Lots of vestigial code was in there.