Beefed up Boss Rewards
Added a small section to the end of boss waves after you defeat a boss that spawns in a lot of mosquitos for the player to catch instead of having them magically added to the player's bank. This adds a visual and unique element to the boss reward instead of just a boring number increase.