Started Work on the Upgrade Timer
Started work on the upgrade timer so the player can see how long they would have the score multiplier upgrade or the better heals upgrade. Added basic functionality of the timer and the progress bar, but have not linked it to the GUI or the rest of the game yet.