Upgrade Visuals Fully Functional
I think I have the upgrade visuals working exactly how I want them to. I have just a few things to add, like maybe something more for the ES and also maybe add the mossies HoT or DoT to the visualizer as well. But for now, you can see the timing of the score multi and better heals as intended.