Heli-Boss now is Functional
The third boss, The Helicopter, now actually can deal damage. Idk if it is balanced or not, but it being 3/5 i wanted it to be able to get some chip damage in and be difficult but probably not lethal. Bosses 4 and 5 will have a lot more damaging attacks and will be more dangerous, but this one is meant to be able to get some chip damage in and finish you off if you entered the fight on low hp.